Fix bugs of create, read, update, and exist functions in vpx_service.
Changed a resource id from service name to vpxid:vipid:serviceName
Added a logic to check if there is a resource which has same service
name with a new resource name in a create function.
UpdateLiveLoadBalancer is a Remove waitForNadcLbServiceUpdateCompletion
and use UpdateLiveLoadBalancer directly.
Use d.Id() in read, updata, and exist functions.