Implicitly import ssh key upon resource create if key already exists
Before creating the SSH key, we first check whether it exists in
SoftLayer. We do this by calculating the fingerprint of the key
locally. Then we look for keys matching that fingerprint (using a
filter). If a matching key is found, we import its details (e.g. id).
We still update the label and notes as specified in the config if
pertinent.